pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiayan-mcp"
version = "0.1.0"
description = "MCP server for Xiayan, a Python-based Markdown formatting tool for WeChat Official Account publishing"
authors = [
    {name = "Xiayan Team", email = "xiayan@example.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["mcp", "markdown", "wechat", "publishing", "xiayan"]

dependencies = [
    "mcp>=1.0.0",
    "markdown>=3.5.0",
    "pygments>=2.16.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-frontmatter>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
xiayan-mcp = "xiayan_mcp.server:main"

[project.urls]
Homepage = "https://github.com/xiayan/xiayan-mcp"
Repository = "https://github.com/xiayan/xiayan-mcp"
Issues = "https://github.com/xiayan/xiayan-mcp/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true